pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dgdm-histopath"
version = "0.1.0"
description = "Dynamic Graph Diffusion Models for Whole-Slide Histopathology Analysis"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "DGDM Team", email = "contact@example.com"}
]
maintainers = [
    {name = "DGDM Team", email = "contact@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
keywords = ["histopathology", "graph neural networks", "diffusion models", "medical AI", "digital pathology"]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torch-geometric>=2.3.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "pillow>=8.3.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
    "pyyaml>=6.0",
    "h5py>=3.1.0",
    "openslide-python>=1.1.2",
    "albumentations>=1.3.0",
    "timm>=0.9.0",
    "transformers>=4.20.0",
    "lightning>=2.0.0",
    "wandb>=0.15.0",
    "tensorboard>=2.8.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "joblib>=1.1.0",
]

[project.optional-dependencies]
viz = [
    "plotly>=5.0.0",
    "bokeh>=2.4.0",
    "holoviews>=1.15.0",
    "umap-learn>=0.5.0",
    "matplotlib-scalebar>=0.8.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "nbsphinx>=0.8.0",
]
clinical = [
    "pydicom>=2.3.0",
    "SimpleITK>=2.2.0",
    "cryptography>=40.0.0",
    "python-jose>=3.3.0",
]
federated = [
    "flower>=1.4.0",
    "opacus>=1.4.0",
    "syft>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/danieleschmidt/dgdm-histopath-lab"
Documentation = "https://dgdm-histopath.readthedocs.io"
Repository = "https://github.com/danieleschmidt/dgdm-histopath-lab"
"Bug Tracker" = "https://github.com/danieleschmidt/dgdm-histopath-lab/issues"

[project.scripts]
dgdm-train = "dgdm_histopath.cli:train"
dgdm-predict = "dgdm_histopath.cli:predict"
dgdm-preprocess = "dgdm_histopath.cli:preprocess"

[tool.setuptools.packages.find]
where = ["."]
include = ["dgdm_histopath*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "openslide.*",
    "albumentations.*",
    "timm.*",
    "torch_geometric.*",
]
ignore_missing_imports = true